requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.13
pandas==2.0.0
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.28.2
pymysql==1.0.3
cryptography==40.0.2
prometheus-client==0.17.0
prometheus-flask-exporter==0.1.2